Fill in template with realistic values
The file is as downloaded from the website, with data filled in for two patients. This can be converted by the website to a "json input" file, which is posted to the normcomp function in andistats, which then gives back a json file with test results for visualisation and tables. At least, that's the plan.